Move ExposedByteArrayOutputStream to io package
This project doesn't have a single root non-domain package, should one be added?
